% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_functions.R
\name{tdt_power_from_ET_ENT}
\alias{tdt_power_from_ET_ENT}
\title{Transmission Disequilibrium Test (TDT) Power from Expected Transmissions and Non-Transmissions}
\usage{
tdt_power_from_ET_ENT(ET, ENT, alpha = 0.05)
}
\arguments{
\item{ET}{Numeric. Expected number of transmissions.}

\item{ENT}{Numeric. Expected number of non-transmissions.}

\item{alpha}{Numeric. Significance level (default = 0.05).}
}
\value{
A list containing:
\item{lambda}{Non-centrality parameter.}
\item{power}{Computed power at the given alpha level.}
\item{ET}{Expected transmissions.}
\item{ENT}{Expected non-transmissions.}
}
\description{
Computes the statistical power of the Transmission Disequilibrium Test (TDT)
given the expected number of transmissions (ET) and non-transmissions (ENT)
under a specified significance level. Implements Equation 1.25 from
\emph{Gordon et al. (2020), Heterogeneity in Statistical Genetics}.
}
\examples{
# Example: compute power for ET = 140 and ENT = 100
tdt_power_from_ET_ENT(ET = 140, ENT = 100, alpha = 0.05)

}
\references{
Gordon, D., Finch, S. J., & Nothnagel, M. (2020).
\emph{Heterogeneity in Statistical Genetics}. Springer Nature.
}
